Moved the disaggregated reports to the reports2019 folder
