+++ api/src/main/java/org/openmrs/module/ugandaemrreports/reports2019/Setup106A1A2019SectionsHC01ToHC15Report.java
@@ -1,4 +1,4 @@
-package org.openmrs.module.ugandaemrreports.reports;
+package org.openmrs.module.ugandaemrreports.reports2019;
 
 import org.openmrs.module.reporting.ReportingConstants;
 import org.openmrs.module.reporting.cohort.definition.BaseObsCohortDefinition;
@@ -20,6 +20,7 @@ import org.openmrs.module.ugandaemrreports.reporting.library.cohort.ARTCohortLib
 import org.openmrs.module.ugandaemrreports.reporting.library.cohort.CommonCohortLibrary;
 import org.openmrs.module.ugandaemrreports.reporting.metadata.Dictionary;
 import org.openmrs.module.ugandaemrreports.reporting.metadata.Metadata;
+import org.openmrs.module.ugandaemrreports.reports.UgandaEMRDataExportManager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
