+++ api/src/main/java/org/openmrs/module/ugandaemrreports/reports2019/Setup106A1A2019SectionsHC16ToHC28Report.java
@@ -1,4 +1,4 @@
-package org.openmrs.module.ugandaemrreports.reports;
+package org.openmrs.module.ugandaemrreports.reports2019;
 
 import org.openmrs.module.reporting.ReportingConstants;
 import org.openmrs.module.reporting.cohort.definition.BaseObsCohortDefinition;
@@ -18,6 +18,7 @@ import org.openmrs.module.ugandaemrreports.library.*;
 import org.openmrs.module.ugandaemrreports.metadata.HIVMetadata;
 import org.openmrs.module.ugandaemrreports.reporting.metadata.Dictionary;
 import org.openmrs.module.ugandaemrreports.reporting.metadata.Metadata;
+import org.openmrs.module.ugandaemrreports.reports.UgandaEMRDataExportManager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
